refactor: move docs link to the separate component
