+++ src/components/pages/doc/docs-link/index.js
@@ -1,0 +1,3 @@
+import DocsLink from './docs-link';
+
+export default DocsLink;
